# Various Jpeg optimizations