… src/ImageSharp/Formats/Jpeg/Components/Block8x8F.cs
… src/ImageSharp/Formats/Jpeg/Components/Decoder/IJpegComponent.cs
… src/ImageSharp/Formats/Jpeg/Components/Decoder/JpegComponent.cs
… src/ImageSharp/Formats/Jpeg/Components/Decoder/JpegComponentExtensions.cs
… src/ImageSharp/Formats/Jpeg/Components/Decoder/JpegComponentPostProcessor.cs
… src/ImageSharp/Formats/Jpeg/JpegDecoderCore.cs
… src/ImageSharp/Image.Decode.cs
… src/ImageSharp/MetaData/Profiles/Exif/ExifReader.cs
… tests/ImageSharp.Benchmarks/Codecs/Jpeg/BlockOperations/Block8x8F_CopyTo1x1.cs
… tests/ImageSharp.Benchmarks/Codecs/Jpeg/BlockOperations/Block8x8F_CopyTo2x2.cs
… tests/ImageSharp.Benchmarks/Codecs/Jpeg/BlockOperations/Block8x8F_DivideRound.cs
… tests/ImageSharp.Benchmarks/Codecs/Jpeg/BlockOperations/Block8x8F_LoadFromInt16.cs
… tests/ImageSharp.Benchmarks/Codecs/Jpeg/BlockOperations/Block8x8F_Round.cs
… tests/ImageSharp.Benchmarks/Codecs/Jpeg/DecodeJpeg.cs
… tests/ImageSharp.Benchmarks/Codecs/Jpeg/DecodeJpegMultiple.cs
… tests/ImageSharp.Benchmarks/Codecs/Jpeg/DecodeJpeg_Aggregate.cs
… tests/ImageSharp.Benchmarks/Codecs/Jpeg/DecodeJpeg_ImageSpecific.cs
… tests/ImageSharp.Benchmarks/Codecs/Jpeg/LoadResizeSave.cs
… tests/ImageSharp.Benchmarks/Codecs/Jpeg/LoadResizeSave_Aggregate.cs
… tests/ImageSharp.Benchmarks/Codecs/Jpeg/LoadResizeSave_ImageSpecific.cs
… tests/ImageSharp.Benchmarks/Codecs/Jpeg/LoadResizeSave_MultiImage.cs
… tests/ImageSharp.Benchmarks/Codecs/MultiImageBenchmarkBase.cs
… tests/ImageSharp.Benchmarks/General/Block8x8F_LoadFromInt16.cs
… tests/ImageSharp.Benchmarks/Samplers/Resize.cs
… tests/ImageSharp.Sandbox46/Program.cs
… tests/ImageSharp.Tests/Common/SimdUtilsTests.cs
… tests/ImageSharp.Tests/ComplexIntegrationTests.cs
… tests/ImageSharp.Tests/Formats/Jpg/Block8x8FTests.CopyToBufferArea.cs
… tests/ImageSharp.Tests/Formats/Jpg/Block8x8FTests.cs
… tests/ImageSharp.Tests/Formats/Jpg/JpegProfilingBenchmarks.cs