#### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/SixLabors/ImageSharp/pulls) open
- [x] I have verified that I am following matches the existing coding patterns and practice as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)

## Summary
This PR contains various optimizations for `JpegDecoder`, resulting in a `~10%` speedup. We are decoding common 420/YCbCr baseline images only 2.4-2.5 slower than System.Drawing now.
I had a hard time squeezing the last drops, we probably can't do much more here before .NET Core 3.0.

## Optimizations
- Replace `JpegComponent.GetBlockReference(x,y)` with `.GetRowSpan(y)`
- Introduce `Image.CreateUninitialized<TPixel>(...)` because decoders do not need pre-initialized image buffers. Introduced in Jpeg, can be used in other decoders as well.
- Optimized `Block8x8` -> `Block8x8F` conversion (`Block8x8F.LoadFrom(...)`) using `Vector.Widen(...)` and `Vector.ConvertToSingle(...)` (.NET Core 2.1 / .NET 4.7.2 only)
- Optimized `Block8x8F.CopyTo(...)` for 2x2 scale cases
- Optimized `ExifReader.ToEnum<T>(...)`

## Chore
- Refactor and group profiling tests
- Add JpegSnoop reports for all images under `TestImages\Input\Jpg`
- `TestImages.Jpeg.BenchmarkSuite`
- Add/refactor benchmarks
- Execute benchmarks add results

## Benchmark results

Only .NET Core 2.1 for now, using the benchmark [DecodeJpeg_ImageSpecific](https://github.com/SixLabors/ImageSharp/blob/af/jpeg-microoptimization/tests/ImageSharp.Benchmarks/Codecs/Jpeg/DecodeJpeg_ImageSpecific.cs#L53) (renamed from `DecodeJpeg`).

#### Before
```
         Method |                                   TestImage |       Mean |      Error |    StdDev | Scaled | ScaledSD |     Gen 0 |    Gen 1 |    Gen 2 |   Allocated |
--------------- |-------------------------------------------- |-----------:|-----------:|----------:|-------:|---------:|----------:|---------:|---------:|------------:|
 System.Drawing |                       Jpg/baseline/Lake.jpg |   6.460 ms |   1.432 ms | 0.0809 ms |   1.00 |     0.00 |   62.5000 |        - |        - |   205.83 KB |
     ImageSharp |                       Jpg/baseline/Lake.jpg |  20.764 ms |   2.747 ms | 0.1552 ms |   3.21 |     0.04 |         - |        - |        - |    19.95 KB |
                |                                             |            |            |           |        |          |           |          |          |             |
 System.Drawing |                Jpg/baseline/jpeg420exif.jpg |  17.156 ms |   5.838 ms | 0.3298 ms |   1.00 |     0.00 |  218.7500 |        - |        - |   757.04 KB |
     ImageSharp |                Jpg/baseline/jpeg420exif.jpg |  46.414 ms |   8.867 ms | 0.5010 ms |   2.71 |     0.05 |         - |        - |        - |    21.92 KB |
                |                                             |            |            |           |        |          |           |          |          |             |
 System.Drawing | Jpg/issues/Issue518-Bad-RST-Progressive.jpg | 428.494 ms | 124.232 ms | 7.0193 ms |   1.00 |     0.00 | 2375.0000 |        - |        - |  7403.76 KB |
     ImageSharp | Jpg/issues/Issue518-Bad-RST-Progressive.jpg | 432.931 ms |  44.287 ms | 2.5023 ms |   1.01 |     0.01 |  125.0000 | 125.0000 | 125.0000 | 35187.05 KB |
                |                                             |            |            |           |        |          |           |          |          |             |
 System.Drawing |       Jpg/issues/issue750-exif-tranform.jpg |  96.404 ms |  45.786 ms | 2.5870 ms |   1.00 |     0.00 | 1750.0000 |        - |        - |  5492.63 KB |
     ImageSharp |       Jpg/issues/issue750-exif-tranform.jpg | 267.698 ms |  37.973 ms | 2.1456 ms |   2.78 |     0.06 |  750.0000 | 750.0000 | 750.0000 | 58833.47 KB |
```

#### After
```
         Method |                                   TestImage |       Mean |      Error |    StdDev | Scaled | ScaledSD |     Gen 0 |    Gen 1 |    Gen 2 |   Allocated |
--------------- |-------------------------------------------- |-----------:|-----------:|----------:|-------:|---------:|----------:|---------:|---------:|------------:|
 System.Drawing |                       Jpg/baseline/Lake.jpg |   6.117 ms |  0.3923 ms | 0.0222 ms |   1.00 |     0.00 |   62.5000 |        - |        - |   205.83 KB |
     ImageSharp |                       Jpg/baseline/Lake.jpg |  18.126 ms |  0.6023 ms | 0.0340 ms |   2.96 |     0.01 |         - |        - |        - |    19.97 KB |
                |                                             |            |            |           |        |          |           |          |          |             |
 System.Drawing |                Jpg/baseline/jpeg420exif.jpg |  17.063 ms |  2.6096 ms | 0.1474 ms |   1.00 |     0.00 |  218.7500 |        - |        - |   757.04 KB |
     ImageSharp |                Jpg/baseline/jpeg420exif.jpg |  41.366 ms |  1.0115 ms | 0.0572 ms |   2.42 |     0.02 |         - |        - |        - |    21.94 KB |
                |                                             |            |            |           |        |          |           |          |          |             |
 System.Drawing | Jpg/issues/Issue518-Bad-RST-Progressive.jpg | 428.282 ms | 94.9163 ms | 5.3629 ms |   1.00 |     0.00 | 2375.0000 |        - |        - |  7403.76 KB |
     ImageSharp | Jpg/issues/Issue518-Bad-RST-Progressive.jpg | 386.698 ms | 33.0065 ms | 1.8649 ms |   0.90 |     0.01 |  125.0000 | 125.0000 | 125.0000 | 35186.97 KB |
                |                                             |            |            |           |        |          |           |          |          |             |
 System.Drawing |       Jpg/issues/issue750-exif-tranform.jpg |  95.192 ms |  3.1762 ms | 0.1795 ms |   1.00 |     0.00 | 1750.0000 |        - |        - |  5492.63 KB |
     ImageSharp |       Jpg/issues/issue750-exif-tranform.jpg | 230.158 ms | 48.8128 ms | 2.7580 ms |   2.42 |     0.02 |  312.5000 | 312.5000 | 312.5000 | 58834.66 KB |
```